requirements.txt
Django==3.1.6
djangorestframework==3.12.2
psycopg2-binary
gunicorn==20.0.4
django-extensions
requests
pyyaml
uritemplate